Update kubernetes python script
