+++ deployment/kubernetes/start.sh
@@ -1,6 +1,7 @@
 #!/bin/bash -e
 
 DIR=$(dirname $(readlink -f "$0"))
+EXT=*.yaml
 
 for i in $(find "$DIR/" -name "*service.yaml"); do
     len=$(echo $DIR | wc -m)
@@ -25,7 +26,7 @@ for i in $(find "$DIR" -name "*deployment.yaml"); do
     done
 done
 
-rm -rf "$DIR/*.yaml"
+rm -rf $DIR/$EXT
 
 yml="$DIR/docker-compose-template.yml"
 test -f "$yml"
